Waiting Info from the Front-End develeper.
Added Comments.
Waiting Info from the Front-End develeper to make additional changes.
There is an idea of getting spark-jobs dynamically from a mongo database. It will be considered in next updates of the orchestrator.

This repository will soon have a Dockerfile to make an image of the orchestrator for easier deployment.