Waiting Info from the Front-End develeper.
Added a Dockerfile to use the Orchestrator in a container.
Waiting Info from the Front-End develeper to make additional changes.
There is an idea of getting spark-jobs dynamically from a mongo database. It will be considered in next updates of the orchestrator.

Soon to be added: A test file to test the Orchestrator and an MD file as a documentation.